new commit

+++ App.py
@@ -1,4 +1,4 @@
-from flask import Flask,render_template,url_for,session,request,redirect
+from flask import Flask,render_template,url_for,session,request,redirect,flash
 import sqlite3
 app = Flask(__name__)
 app.secret_key = "ahgakjg89"
@@ -39,8 +39,10 @@ def bookData():
             try:
                 cur = con.cursor()
                 cur.execute("INSERT INTO book_details(book_id,book_name,book_qty)VALUES(?,?,?)",(bookId,bookName,bookQty))
+                flash("details added successfully!")
                 con.commit()
             except:
+                flash("details not added!")
                 con.rollback()
             finally:
                 return redirect(url_for("viewBooks"))
@@ -64,9 +66,11 @@ def updateBookData():
             try:
                 cur = con.cursor()
                 cur.execute("UPDATE book_details SET book_name = ?,book_qty = ? WHERE book_id = ?",(bookName,bookQty,bookId))
+                flash("Details Updated Successfully!")
                 con.commit()
             except:
                 cur.rollback()
+                flash("Details not Updated!")
             finally:
                 return redirect(url_for('viewBooks'))
   
@@ -79,9 +83,11 @@ def deleteBookData():
             try:
                 cur = con.cursor()
                 cur.execute("DELETE FROM book_details WHERE book_id = ?",(bookId))
+                flash("Details Deleted Successfully!")
                 con.commit()
             except:
                 con.rollback()
+                flash("Details not Deleted!")
             finally:
                 return redirect(url_for("viewBooks"))
             
@@ -112,9 +118,12 @@ def memberData():
             try:
                 cur = con.cursor()
                 cur.execute("INSERT INTO member_details(member_id,member_name,email)VALUES(?,?,?)",(memberId,memberName,str(email)))
+                flash("Details Added successfully!")
                 con.commit()
             except:
                 con.rollback()
+                flash("Details Not Added!")
+
             finally:
                 return redirect(url_for("viewMembers"))
             
@@ -132,9 +141,11 @@ def updateMemberData():
             try:
                 cur = con.cursor()
                 cur.execute("UPDATE member_details SET member_name = ?,email = ? WHERE member_id = ?",(memberName,email,memberId))
+                flash("Details Updated Successfully!")
                 con.commit()
             except:
                 cur.rollback()
+                flash("Details Not Updated!")
 
             finally:
                 return redirect(url_for('viewMembers'))
@@ -152,9 +163,12 @@ def deleteMemberData():
             try:
                 cur = con.cursor()
                 cur.execute("DELETE FROM member_details WHERE member_id = ?",(memberId))
+                flash("Details Deleted Successfully!")
                 con.commit()
             except:
                 con.rollback()
+                flash("Details not Deleted!")
+
             finally:
                 return redirect(url_for("viewMembers"))
             
@@ -173,7 +187,6 @@ def issueBook():
 
 @app.route("/issueBookMember",methods=['GET','POST'])
 def issueBookMember():
-    msg = ""
     if request.method == "POST":
         bookId = request.form["bookId"]
         memberId = request.form["memberId"]
@@ -187,22 +200,21 @@ def issueBookMember():
                     cur.execute("SELECT book_qty FROM book_details WHERE book_id = ?",(bookId))
                     qty = cur.fetchone()[0]
                     if qty == 0:
-                        msg = "Out of Stock!"
+                        flash("Out of Stock!")
                     elif (qty-int(bookQty)) < 0:
-                        msg = f"We Have Only {qty} Books!"
+                        flash(f"We Have Only {qty} Books!")
                     else:
                         cur.execute("SELECT member_id FROM member_details WHERE member_id = ?",(memberId))
                         if cur.fetchone()[0]:
-                            msg = f"{bookQty} of {bookId} books issued to {memberId} Successfully!"
+                            flash(f"{bookQty} of {bookId} books issued to {memberId} Successfully!")
                             cur.execute("SELECT book_qty FROM book_details WHERE book_id = ?",(bookId))
                             qty = cur.fetchone()[0]
                             cur.execute("UPDATE book_details SET book_qty = ? WHERE book_id = ?",((qty-int(bookQty)),bookId))
                             cur.execute("INSERT INTO issue_details(book_id,member_id,due_date,book_qty)VALUES(?,?,?,?)",(bookId,memberId,dueDate,bookQty))
                         else:
-                            msg = f"{memberId} is not a member of this library!"
+                            flash(f"{memberId} is not a member of this library!")
             except:   
-                msg = "Book id is not exist!"
-                print(msg)
+                 flash("Book id is not exist!")
             con.commit()
     return redirect(url_for("viewIssued"))
 
@@ -240,9 +252,12 @@ def deleteIssueData():
             try:
                 cur = con.cursor()
                 cur.execute("DELETE FROM issue_details WHERE member_id = ? and book_id = ?",(memberId,bookId))
+                flash("Details Deleted Successfully!")
                 con.commit()
             except:
                 con.rollback()
+                flash("Details not Deleted!")
+
             finally:
                 return redirect(url_for("viewIssued"))
 
@@ -252,7 +267,6 @@ def issueReturn():
 
 @app.route("/returnBook",methods=["GET","POST"])
 def returnBook():
-    msg = ""
     if request.method == "POST":
         bookId = request.form["bookId"]
         memberId = request.form["memberid"]
@@ -266,32 +280,38 @@ def returnBook():
                     qty = cur.fetchone()[0]
                     cur.execute("SELECT member_id FROM member_details WHERE member_id = ?",(memberId))
                     try:
-                        if cur.fetchone()[0]:
-                            msg = f"{bookQty} of {bookId} books returned from {memberId} Successfully!"
-                            cur.execute("SELECT book_qty FROM book_details WHERE book_id = ?",(bookId))
-                            qty = cur.fetchone()[0]
-                            cur.execute("UPDATE book_details SET book_qty = ? WHERE book_id = ?",((qty+int(bookQty)),bookId))
-                            cur.execute("SELECT book_qty FROM issue_details WHERE book_id = ? and member_id = ?",(bookId,memberId))
+                        if cur.fetchone():
+                            cur.execute("SELECT member_id FROM issue_details WHERE member_id = ?",(memberId))                        
                             try:
-                                qtyIssue = cur.fetchone()[0]
-                                if qtyIssue < int(bookQty):
-                                    msg=f"You Didn't have {bookQty} books to return!"
-                                else:
-                                    finalData = qtyIssue - int(bookQty)
-                                    cur.execute("UPDATE issue_details SET book_qty = ? WHERE book_id = ? and member_id = ?",(finalData,bookId,memberId))
+                                if cur.fetchone()[0]:
+                                    flash(f"{bookQty} of {bookId} books returned from {memberId} Successfully!")
+                                    cur.execute("SELECT book_qty FROM book_details WHERE book_id = ?",(bookId))
+                                    qty = cur.fetchone()[0]
+                                    cur.execute("UPDATE book_details SET book_qty = ? WHERE book_id = ?",((qty+int(bookQty)),bookId))
+                                    flash("Details Updated Successfully!")
                                     cur.execute("SELECT book_qty FROM issue_details WHERE book_id = ? and member_id = ?",(bookId,memberId))
-                                    qtyIssue = cur.fetchone()[0]
-                                    if qtyIssue <= 0:
-                                        msg="You Didn't have any books to return!"
-                                        cur.execute("DELETE FROM issue_details WHERE book_id = ? and member_id = ?",(bookId,memberId))
-                            except:
-                                msg = "book id not exist!"
+                                    try:
+                                        qtyIssue = cur.fetchone()[0]
+                                        if qtyIssue < int(bookQty):
+                                            flash(f"You Didn't have {bookQty} books to return!")
+                                        else:
+                                            finalData = qtyIssue - int(bookQty)
+                                            cur.execute("UPDATE issue_details SET book_qty = ? WHERE book_id = ? and member_id = ?",(finalData,bookId,memberId))
+                                            flash("Details Updated Successfully!")
+                                            cur.execute("SELECT book_qty FROM issue_details WHERE book_id = ? and member_id = ?",(bookId,memberId))
 
+                                            qtyIssue = cur.fetchone()[0]
+                                            if qtyIssue <= 0:
+                                                flash("You Didn't have any books to return!")
+                                                cur.execute("DELETE FROM issue_details WHERE book_id = ? and member_id = ?",(bookId,memberId))
+                                    except:
+                                        flash("book id not exist!")
+                            except:
+                                flash(f"member {memberId} has no books to return!")
                     except:
-                        msg = f"{memberId} is not a member of this library!"
+                        flash(f"member {memberId} is not a member of this library!")
             except:   
-                msg = "Book id is not exist!"
-                print(msg)
+                flash("Book id is not exist!")
             con.commit()
     return redirect(url_for("viewBooks"))
